Added first part

+++ Exercises/MetricsAgent/DAL/TimeSpanHandler.cs
@@ -9,7 +9,7 @@ namespace MetricsAgent.DAL
 {
     public class TimeSpanHandler : SqlMapper.TypeHandler<TimeSpan>
     {
-        public override TimeSpan Parse(object value) => TimeSpan.FromSeconds((int)value);
+        public override TimeSpan Parse(object value) => TimeSpan.FromSeconds((long)value);
         public override void SetValue(IDbDataParameter parameter, TimeSpan value) => parameter.Value = value;
     }
 }
